+++ apps/hubble/src/hubble.ts
@@ -65,7 +65,7 @@ import { AddrInfo } from "@chainsafe/libp2p-gossipsub/types";
 import { CheckIncomingPortsJobScheduler } from "./storage/jobs/checkIncomingPortsJob.js";
 import { NetworkConfig, applyNetworkConfig, fetchNetworkConfig } from "./network/utils/networkConfig.js";
 import { UpdateNetworkConfigJobScheduler } from "./storage/jobs/updateNetworkConfigJob.js";
-import { statsd } from "./utils/statsd.js";
+import { statsd, StatsDInitParams } from "./utils/statsd.js";
 import {
   getDbSchemaVersion,
   LATEST_DB_SCHEMA_VERSION,
@@ -174,6 +174,9 @@ export interface HubOptions {
 
   /** Rank RPCs and use the ones with best stability and latency */
   rankRpcs?: boolean;
+
+  /** StatsD parameters */
+  statsdParams?: StatsDInitParams | undefined;
 
   /** ETH mainnet RPC URL(s) */
   ethMainnetRpcUrl?: string;
@@ -692,6 +695,7 @@ export class Hub implements HubInterface {
       applicationScoreCap: this.options.applicationScoreCap,
       strictNoSign: this.strictNoSign,
       connectToDbPeers: this.options.connectToDbPeers,
+      statsdParams: this.options.statsdParams,
     });
 
     await this.registerEventHandlers();
@@ -1017,7 +1021,7 @@ export class Hub implements HubInterface {
     if (gossipMessage.timestamp) {
       // If message is older than seenTTL, we will try to merge it, but report it as invalid so it doesn't
       // propogate across the network
-      const cutOffTime = getFarcasterTime().unwrapOr(0) - GOSSIP_SEEN_TTL;
+      const cutOffTime = getFarcasterTime().unwrapOr(0) - GOSSIP_SEEN_TTL / 1000;
 
       if (gossipMessage.timestamp < cutOffTime) {
         await this.gossipNode.reportValid(msgId, peerIdFromString(source.toString()).toBytes(), false);
@@ -1049,6 +1053,10 @@ export class Hub implements HubInterface {
         return err(new HubError("unavailable", "Sync queue is full"));
       }
 
+      const currentTime = getFarcasterTime().unwrapOr(0);
+      const messageFirstGossipedTime = gossipMessage.timestamp ?? 0;
+      const gossipMessageDelay = currentTime - messageFirstGossipedTime;
+
       // Merge the message
       const result = await this.submitMessage(message, "gossip");
       if (result.isOk()) {
@@ -1056,12 +1064,24 @@ export class Hub implements HubInterface {
           await this.gossipNode.reportValid(msgId, peerIdFromString(source.toString()).toBytes(), true);
         }
       } else {
+        const tags: { [key: string]: string } = {
+          valid: reportedAsInvalid ? "false" : "true",
+          error_code: result.error.errCode,
+          message_type: messageTypeToName(message.data?.type),
+        };
+
+        statsd().increment("gossip.message_failure", 1, tags);
         log.info(
           {
             errCode: result.error.errCode,
+            errMsg: result.error.message,
             peerId: source.toString(),
             origin: peerIdResult.value,
             hash: bytesToHexString(message.hash).unwrapOr(""),
+            fid: message.data?.fid,
+            type: message.data?.type,
+            gossipDelay: gossipMessageDelay,
+            valid: !reportedAsInvalid,
             msgId,
           },
           "Received bad gossip message from peer",
@@ -1071,20 +1091,14 @@ export class Hub implements HubInterface {
         }
       }
 
-      const currentTime = getFarcasterTime().unwrapOr(0);
-      const messageCreatedTime = message.data?.timestamp ?? 0;
-      // The message time is user provided, so, while not ideal, it's still good enough to use most of the time
-      if (currentTime > 0 && messageCreatedTime > 0 && currentTime > messageCreatedTime) {
-        const diff = currentTime - messageCreatedTime;
-        statsd().timing("gossip.message_delay", diff);
-        const mergeResult = result.isOk() ? "success" : "failure";
-        statsd().timing(`gossip.message_delay.${mergeResult}`, diff);
-      }
+      statsd().timing("gossip.message_delay", gossipMessageDelay);
+      const mergeResult = result.isOk() ? "success" : "failure";
+      statsd().timing(`gossip.message_delay.${mergeResult}`, gossipMessageDelay);
 
       return result.map(() => undefined);
     } else if (gossipMessage.contactInfoContent) {
       const result = await this.handleContactInfo(peerIdResult.value, gossipMessage.contactInfoContent);
-      this.gossipNode.reportValid(msgId, peerIdFromString(source.toString()).toBytes(), result);
+      await this.gossipNode.reportValid(msgId, peerIdFromString(source.toString()).toBytes(), result);
       return ok(undefined);
     } else {
       return err(new HubError("bad_request.invalid_param", "invalid message type"));
@@ -1230,7 +1244,8 @@ export class Hub implements HubInterface {
     // TODO remove this once all peers are updated past 1.6.4
     if (message.timestamp === 0) return false;
 
-    return true;
+    // Return if peer contact was newer than existing contact (prevents old gossip messages from being forwarded)
+    return result.isOk();
   }
 
   /** Since we don't know if the peer is using SSL or not, we'll attempt to get the SSL version,
@@ -1361,8 +1376,9 @@ export class Hub implements HubInterface {
   /* -------------------------------------------------------------------------- */
 
   async submitMessage(submittedMessage: Message, source?: HubSubmitSource): HubAsyncResult<number> {
-    // If this is a dup, don't bother processing it
-    if (await isMessageInDB(this.rocksDB, submittedMessage)) {
+    // If this is a dup, don't bother processing it. Only do this for gossip messages since rpc messages
+    // haven't been validated yet
+    if (source === "gossip" && (await isMessageInDB(this.rocksDB, submittedMessage))) {
       log.debug({ source }, "submitMessage rejected: Message already exists");
       return err(new HubError("bad_request.duplicate", "message has already been merged"));
     }
@@ -1403,7 +1419,12 @@ export class Hub implements HubInterface {
       },
       (e) => {
         logMessage.warn({ errCode: e.errCode, source }, `submitMessage error: ${e.message}`);
-        statsd().increment(`submit_message.error.${source}.${e.errCode}`);
+        const tags: { [key: string]: string } = {
+          error_code: e.errCode,
+          message_type: type,
+          source: source ?? "unknown-source",
+        };
+        statsd().increment("submit_message.error", 1, tags);
       },
     );