+++ apps/hubble/src/network/sync/syncEngine.ts
@@ -431,15 +431,27 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
     const existingPeerInfo = this.getContactInfoForPeerId(peerId.toString());
     if (existingPeerInfo) {
       if (contactInfo.timestamp > existingPeerInfo.contactInfo.timestamp) {
-        log.debug({ peerInfo: existingPeerInfo }, "Updating peer with latest contactInfo");
         this.currentHubPeerContacts.set(peerId.toString(), { peerId, contactInfo });
       }
       return err(new HubError("bad_request.duplicate", "peer already exists"));
     } else {
-      log.info({ peerInfo: contactInfo, connectedPeers: this.getPeerCount() }, "New Peer ContactInfo");
       this.currentHubPeerContacts.set(peerId.toString(), { peerId, contactInfo });
-      return ok(undefined);
     }
+    log.info(
+      {
+        peerInfo: contactInfo,
+        theirMessages: contactInfo.count,
+        peerNetwork: contactInfo.network,
+        peerVersion: contactInfo.hubVersion,
+        peerAppVersion: contactInfo.appVersion,
+        connectedPeers: this.getPeerCount(),
+        peerId: peerId.toString(),
+        isNew: !!existingPeerInfo,
+        gossipDelay: (Date.now() - contactInfo.timestamp) / 1000,
+      },
+      "Updated Peer ContactInfo",
+    );
+    return ok(undefined);
   }
 
   public removeContactInfoForPeerId(peerId: string) {