+++ apps/hubble/src/storage/db/onChainEvent.ts
@@ -1,7 +1,7 @@
 import RocksDB, { Iterator, Transaction } from "./rocksdb.js";
 import { HubError, OnChainEvent, OnChainEventType } from "@farcaster/hub-nodejs";
 import { OnChainEventPostfix, RootPrefix } from "./types.js";
-import { getPageIteratorByPrefix, makeFidKey } from "./message.js";
+import { getPageIteratorOptsByPrefix, makeFidKey } from "./message.js";
 import { PAGE_SIZE_MAX, PageOptions } from "../stores/types.js";
 import { ResultAsync } from "neverthrow";
 
@@ -32,9 +32,7 @@ export const makeOnChainEventPrimaryKey = (
   logIndex: number,
 ): Buffer => {
   return Buffer.concat([
-    Buffer.from([RootPrefix.OnChainEvent]),
-    Buffer.from([OnChainEventPostfix.OnChainEvents]),
-    Buffer.from([type]),
+    Buffer.from([RootPrefix.OnChainEvent, OnChainEventPostfix.OnChainEvents, type]),
     makeFidKey(fid),
     makeBlockNumberKey(blockNumber),
     makeLogIndexKey(logIndex),
@@ -43,35 +41,25 @@ export const makeOnChainEventPrimaryKey = (
 
 export const makeSignerOnChainEventBySignerKey = (fid: number, signer: Uint8Array): Buffer => {
   return Buffer.concat([
-    Buffer.from([RootPrefix.OnChainEvent]),
-    Buffer.from([OnChainEventPostfix.SignerByFid]),
+    Buffer.from([RootPrefix.OnChainEvent, OnChainEventPostfix.SignerByFid]),
     makeFidKey(fid),
     Buffer.from(signer),
   ]);
 };
 
 export const makeIdRegisterEventByFidKey = (fid: number): Buffer => {
-  return Buffer.concat([
-    Buffer.from([RootPrefix.OnChainEvent]),
-    Buffer.from([OnChainEventPostfix.IdRegisterByFid]),
-    makeFidKey(fid),
-  ]);
+  return Buffer.concat([Buffer.from([RootPrefix.OnChainEvent, OnChainEventPostfix.IdRegisterByFid]), makeFidKey(fid)]);
 };
 
 export const makeIdRegisterEventByCustodyKey = (custodyAddress: Uint8Array): Buffer => {
   return Buffer.concat([
-    Buffer.from([RootPrefix.OnChainEvent]),
-    Buffer.from([OnChainEventPostfix.IdRegisterByCustodyAddress]),
+    Buffer.from([RootPrefix.OnChainEvent, OnChainEventPostfix.IdRegisterByCustodyAddress]),
     Buffer.from(custodyAddress),
   ]);
 };
 
 export const makeOnChainEventIteratorPrefix = (type: OnChainEventType, fid?: number): Buffer => {
-  let prefix = Buffer.concat([
-    Buffer.from([RootPrefix.OnChainEvent]),
-    Buffer.from([OnChainEventPostfix.OnChainEvents]),
-    Buffer.from([type]),
-  ]);
+  let prefix = Buffer.concat([Buffer.from([RootPrefix.OnChainEvent, OnChainEventPostfix.OnChainEvents, type])]);
   if (fid) {
     prefix = Buffer.concat([prefix, makeFidKey(fid)]);
   }
@@ -153,36 +141,34 @@ export const getOnChainEventsPageByPrefix = async <T extends OnChainEvent>(
   events: T[];
   nextPageToken: Uint8Array | undefined;
 }> => {
-  const iterator = getPageIteratorByPrefix(db, prefix, pageOptions);
+  const iteratorOpts = getPageIteratorOptsByPrefix(db, prefix, pageOptions);
 
   const limit = pageOptions.pageSize || PAGE_SIZE_MAX;
 
   const eventsPrimaryKeys: Buffer[] = [];
 
-  const getNextIteratorRecord = async (iterator: Iterator): Promise<[Buffer, Buffer]> => {
-    const [key, value] = await iterator.next();
-    return [key as Buffer, value as Buffer];
-  };
-
-  let iteratorFinished = false;
+  let iteratorFinished = true;
   let lastPageToken: Uint8Array | undefined;
-  do {
-    const result = await ResultAsync.fromPromise(getNextIteratorRecord(iterator), (e) => e as HubError);
-    if (result.isErr()) {
-      iteratorFinished = true;
-      break;
-    }
 
-    const [key, value] = result.value;
+  await db.forEachIteratorByOpts(iteratorOpts, async (key, value) => {
+    if (!key) {
+      return false; // skip
+    }
     lastPageToken = Uint8Array.from(key.subarray(prefix.length));
     if (value) {
       eventsPrimaryKeys.push(value);
+
+      if (eventsPrimaryKeys.length >= limit) {
+        iteratorFinished = false;
+        return true; // stop
+      }
     }
-  } while (eventsPrimaryKeys.length < limit);
+
+    return false;
+  });
 
   const events = (await getManyOnChainEvents<T>(db, eventsPrimaryKeys)).filter(filter);
 
-  await iterator.end();
   if (!iteratorFinished) {
     return { events: events, nextPageToken: lastPageToken };
   } else {