+++ apps/hubble/src/storage/stores/store.ts
@@ -13,7 +13,7 @@ import {
   getManyMessages,
   getMessage,
   getMessagesPageByPrefix,
-  getPageIteratorByPrefix,
+  getPageIteratorOptsByPrefix,
   makeMessagePrimaryKey,
   makeTsHash,
   messageDecode,
@@ -68,30 +68,22 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
   abstract findMergeAddConflicts(message: TAdd): HubAsyncResult<void>;
   abstract findMergeRemoveConflicts(message: TRemove): HubAsyncResult<void>;
 
-  async validateAdd(add: TAdd): HubAsyncResult<void> {
+  async validateAdd(add: TAdd): HubAsyncResult<Uint8Array> {
     if (!add.data) {
       return err(new HubError("bad_request.invalid_param", "data null"));
     }
 
     const tsHash = makeTsHash(add.data.timestamp, add.hash);
-    if (tsHash.isErr()) {
-      return err(tsHash.error);
-    }
-
-    return ok(undefined);
+    return tsHash;
   }
 
-  async validateRemove(remove: TRemove): HubAsyncResult<void> {
+  async validateRemove(remove: TRemove): HubAsyncResult<Uint8Array> {
     if (!remove.data) {
       return err(new HubError("bad_request.invalid_param", "data null"));
     }
 
     const tsHash = makeTsHash(remove.data.timestamp, remove.hash);
-    if (tsHash.isErr()) {
-      return err(tsHash.error);
-    }
-
-    return ok(undefined);
+    return tsHash;
   }
 
   async buildSecondaryIndices(_txn: Transaction, _add: TAdd): HubAsyncResult<void> {
@@ -106,7 +98,7 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
     this._eventHandler = eventHandler;
     this._pruneSizeLimit = options.pruneSizeLimit ?? this.PRUNE_SIZE_LIMIT_DEFAULT;
     this._pruneTimeLimit = options.pruneTimeLimit ?? this.PRUNE_TIME_LIMIT_DEFAULT;
-    this._mergeLock = new AsyncLock();
+    this._mergeLock = new AsyncLock({ timeout: MERGE_TIMEOUT_DEFAULT });
   }
 
   /** Looks up TAdd message by tsHash */
@@ -189,32 +181,28 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
 
     return (
       this._mergeLock
-        .acquire(
-          message.data.fid.toString(),
-          async () => {
-            const prunableResult = await this._eventHandler.isPrunable(
-              // biome-ignore lint/suspicious/noExplicitAny: legacy code, avoid using ignore for new code
-              message as any,
-              this._postfix,
-              this.pruneSizeLimit,
-              this.pruneTimeLimit,
-            );
-            if (prunableResult.isErr()) {
-              throw prunableResult.error;
-            } else if (prunableResult.value) {
-              throw new HubError("bad_request.prunable", "message would be pruned");
-            }
+        .acquire(message.data.fid.toString(), async () => {
+          const prunableResult = await this._eventHandler.isPrunable(
+            // biome-ignore lint/suspicious/noExplicitAny: legacy code, avoid using ignore for new code
+            message as any,
+            this._postfix,
+            this.pruneSizeLimit,
+            this.pruneTimeLimit,
+          );
+          if (prunableResult.isErr()) {
+            throw prunableResult.error;
+          } else if (prunableResult.value) {
+            throw new HubError("bad_request.prunable", "message would be pruned");
+          }
 
-            if (this._isAddType(message)) {
-              return this.mergeAdd(message);
-            } else if (this._isRemoveType?.(message)) {
-              return this.mergeRemove(message);
-            } else {
-              throw new HubError("bad_request.validation_failure", "invalid message type");
-            }
-          },
-          { timeout: MERGE_TIMEOUT_DEFAULT },
-        )
+          if (this._isAddType(message)) {
+            return this.mergeAdd(message);
+          } else if (this._isRemoveType?.(message)) {
+            return this.mergeRemove(message);
+          } else {
+            throw new HubError("bad_request.validation_failure", "invalid message type");
+          }
+        })
         // biome-ignore lint/suspicious/noExplicitAny: legacy code, avoid using ignore for new code
         .catch((e: any) => {
           throw isHubError(e) ? e : new HubError("unavailable.storage_failure", "merge timed out");
@@ -363,38 +351,36 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
   }
 
   protected async getBySecondaryIndex(prefix: Buffer, pageOptions: PageOptions = {}): Promise<MessagesPage<TAdd>> {
-    const iterator = getPageIteratorByPrefix(this._db, prefix, pageOptions);
+    const iteratorOpts = getPageIteratorOptsByPrefix(this._db, prefix, pageOptions);
 
     const limit = pageOptions.pageSize || PAGE_SIZE_MAX;
 
     const messageKeys: Buffer[] = [];
 
-    // Custom method to retrieve message key from key
-    const getNextIteratorRecord = async (iterator: Iterator): Promise<[Buffer, Buffer]> => {
-      const [key] = await iterator.next();
+    let iteratorFinished = true;
+    let lastPageToken: Uint8Array | undefined;
+
+    await this._db.forEachIteratorByOpts(iteratorOpts, (key) => {
+      if (!key) {
+        return false; // continue
+      }
       const fid = Number((key as Buffer).readUint32BE(prefix.length + TSHASH_LENGTH));
       const tsHash = Uint8Array.from(key as Buffer).subarray(prefix.length, prefix.length + TSHASH_LENGTH);
-      const messagePrimaryKey = makeMessagePrimaryKey(fid, this._postfix, tsHash);
-      return [key as Buffer, messagePrimaryKey];
-    };
+      const messageKey = makeMessagePrimaryKey(fid, this._postfix, tsHash);
 
-    let iteratorFinished = false;
-    let lastPageToken: Uint8Array | undefined;
-    do {
-      const result = await ResultAsync.fromPromise(getNextIteratorRecord(iterator), (e) => e as HubError);
-      if (result.isErr()) {
-        iteratorFinished = true;
-        break;
-      }
-
-      const [key, messageKey] = result.value;
       lastPageToken = Uint8Array.from(key.subarray(prefix.length));
       messageKeys.push(messageKey);
-    } while (messageKeys.length < limit);
+
+      if (messageKeys.length >= limit) {
+        iteratorFinished = false;
+        return true; // stop
+      }
+
+      return false; // continue
+    });
 
     const messages = await getManyMessages<TAdd>(this._db, messageKeys);
 
-    await iterator.end(); // clear iterator if it has not finished
     if (!iteratorFinished) {
       return { messages, nextPageToken: lastPageToken };
     } else {
@@ -487,6 +473,7 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
     if (validateResult.isErr()) {
       return err(validateResult.error);
     }
+    const tsHash = validateResult.value;
 
     const checkResult = await (this._isAddType(message)
       ? this.findMergeAddConflicts(message)
@@ -494,13 +481,6 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
 
     if (checkResult.isErr()) {
       return err(checkResult.error);
-    }
-
-    // biome-ignore lint/style/noNonNullAssertion: legacy code, avoid using ignore for new code
-    const tsHash = makeTsHash(message.data!.timestamp, message.hash);
-
-    if (tsHash.isErr()) {
-      throw tsHash.error;
     }
 
     const conflicts: (TAdd | TRemove)[] = [];
@@ -518,7 +498,7 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
           new Uint8Array(removeTsHash.value),
           // biome-ignore lint/style/noNonNullAssertion: legacy code, avoid using ignore for new code
           message.data!.type,
-          tsHash.value,
+          tsHash,
         );
         if (removeCompare > 0) {
           return err(new HubError("bad_request.conflict", "message conflicts with a more recent remove"));
@@ -549,7 +529,7 @@ export abstract class Store<TAdd extends Message, TRemove extends Message> {
         new Uint8Array(addTsHash.value),
         // biome-ignore lint/style/noNonNullAssertion: legacy code, avoid using ignore for new code
         message.data!.type,
-        tsHash.value,
+        tsHash,
       );
       if (addCompare > 0) {
         return err(new HubError("bad_request.conflict", "message conflicts with a more recent add"));