+++ apps/hubble/src/utils/statsd.ts
@@ -1,4 +1,4 @@
-import { StatsD } from "hot-shots";
+import { StatsD } from "@figma/hot-shots";
 import { logger } from "./logger.js";
 
 const log = logger.child({ module: "statsd" });
@@ -30,7 +30,7 @@ export function statsd(): StatsD {
 
 let statsdInitialization: StatsDInitParams | undefined;
 
-export function getStatusdInitialization(): StatsDInitParams | undefined {
+export function getStatsdInitialization(): StatsDInitParams | undefined {
   return statsdInitialization;
 }