+++ packages/core/src/validations.ts
@@ -733,6 +733,8 @@ export const validateVerificationRemoveBody = (
   switch (body.protocol) {
     case protobufs.Protocol.ETHEREUM:
       return validateEthAddress(body.address).map(() => body);
+    case protobufs.Protocol.SOLANA:
+      return validateSolAddress(body.address).map(() => body);
     default:
       return err(new HubError("bad_request.validation_failure", "invalid verification protocol"));
   }
@@ -783,6 +785,9 @@ export const validateFrameActionBody = (body: protobufs.FrameActionBody): HubRes
   }
   if (validateBytesAsString(body.inputText, 256).isErr()) {
     return err(new HubError("bad_request.validation_failure", "invalid input text"));
+  }
+  if (validateBytesAsString(body.state, 4096).isErr()) {
+    return err(new HubError("bad_request.validation_failure", "invalid state"));
   }
 
   if (body.castId !== undefined) {