+++ apps/replicator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @farcaster/replicator
 
+## 0.3.2
+
+### Patch Changes
+
+- addf097c: chore: update @farcaster/hub-nodejs
+
+## 0.3.1
+
+### Patch Changes
+
+- a8b7dfcb: fix: Replace hot-shots with @figma/hot-shots
+- Updated dependencies [5703d339]
+  - @farcaster/hub-nodejs@0.11.2
+
 ## 0.3.0
 
 ### Minor Changes